+++ src/main/java/duke/DukeException.java
@@ -1,0 +1,9 @@
+package duke;
+
+public class DukeException extends Exception {
+    public String message;
+
+    public DukeException(String message) {
+        this.message = message;
+    }
+}
