+++ src/main/java/duke/TaskNotExist.java
@@ -1,0 +1,7 @@
+package duke;
+
+class TaskNotExist extends DukeException {
+    public TaskNotExist() {
+        super("Invalid value, there isn't this much tasks in the list :-( ");
+    }
+}
