+++ src/main/java/duke/ToDo.java
@@ -1,0 +1,15 @@
+package duke;
+
+class ToDo extends Task {
+    String icon = "[T]";
+
+    public ToDo(String details) {
+        super(details);
+
+    }
+
+    @Override
+    public String toString() {
+        return icon + super.toString();
+    }
+}
